RestoranApp

+++ model/Jelo.py
@@ -1,5 +1,4 @@
-from . import Base
-
+from __init__ import Base
 from sqlalchemy import *
 from sqlalchemy.orm import relationship
 
@@ -7,5 +6,5 @@ class Food (Base):
     __tablename__ = "hrana"
     ID_food = Column(Integer, primary_key = True)
     naziv = Column(String(50))
-    cijena =Column(Integer(50))
+    cijena =Column(Integer)
