+++ model/Pice.py
@@ -1,4 +1,4 @@
-from . import Base
+from __init__ import Base
 
 from sqlalchemy import *
 from sqlalchemy.orm import relationship
@@ -7,5 +7,5 @@ class Drinks (Base):
     __tablename__ = "drinks"
     ID_drink = Column(Integer, primary_key = True)
     naziv = Column(String(50))
-    cijena =Column(Integer(50))
+    cijena =Column(Integer)
